Drug Tariff Master/requirements.txt
# Drug Tariff Master - Dependencies
# Core libraries
requests==2.31.0
lxml==4.9.3
xmlschema==2.5.1
apscheduler==3.10.4
python-dotenv==1.0.0

# Web interface
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Database
# Using SQLite which is part of the Python standard library

# Development and testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0 